ライブラリの構成を変更した
