+++ lib/twitter/bot.rb
@@ -1,11 +1,12 @@
 require 'twitter'
 
-module CodeCandy
+module Twitter
   class Bot
     class << self
       def tweet(title)
         begin
           text = "【新作問題】\nCodeCandyに新しい問題を追加しました。\nタイトル: #{title}"
+          # puts text
           client.update(text)
         rescue => e
           puts "== TWEET ERROR =="
